remove data-bind attribute
